requirements.txt
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dash==4.0.0
et_xmlfile==2.0.0
Flask==3.1.3
google-auth==2.48.0
google-auth-oauthlib==1.2.4
greenlet==3.3.2
gspread==6.2.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.16.0
nest-asyncio==1.6.0
numpy==2.4.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plotly==6.5.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
retrying==1.4.2
rsa==4.9.1
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
zipp==3.23.0
